+++ nextgisweb/auth/nodepkg/user-widget/UserWidget.js
@@ -1,20 +1,24 @@
 import { Alert } from "@nextgisweb/gui/antd";
 import { ContentBox, LoadingWrapper } from "@nextgisweb/gui/component";
-import { KeynameTextBox, LanguageSelect } from "@nextgisweb/gui/fields-form";
+import {
+    KeynameTextBox,
+    LanguageSelect,
+    Password,
+} from "@nextgisweb/gui/fields-form";
 import { ModelForm } from "@nextgisweb/gui/model-form";
 import { useRouteGet } from "@nextgisweb/pyramid/hook/useRouteGet";
-import { route } from "@nextgisweb/pyramid/api";
 import i18n from "@nextgisweb/pyramid/i18n!auth";
 import { PropTypes } from "prop-types";
 import { useMemo } from "react";
 import { PrincipalMemberSelect } from "../field";
 import { default as oauth, makeTeamManageButton } from "../oauth";
 import getMessages from "../userMessages";
+import { UserWidgetPassword } from "./UserWidgetPassword";
 
 export function UserWidget({ id }) {
-    const { data: group, isLoading } = useRouteGet({
-        name: "auth.group.collection",
-    });
+    const { data: group, isLoading } = useRouteGet("auth.group.collection");
+
+    const isNewUser = useMemo(() => id === undefined, [id]);
 
     const fields = useMemo(() => {
         const fields_ = [];
@@ -35,19 +39,15 @@ export function UserWidget({ id }) {
                 {
                     name: "password",
                     label: i18n.gettext("Password"),
-                    widget: "password",
-                    autoComplete: 'new-password',
-                    // required only when creating a new user
-                    required: id === undefined,
-                    placeholder:
-                        id !== undefined
-                            ? i18n.gettext("Enter new password here")
-                            : "",
+                    widget: isNewUser ? Password : UserWidgetPassword,
+                    required: true,
+                    autoComplete: "new-password",
+                    placeholder: i18n.gettext("Enter new password here"),
                 },
             ]
         );
 
-        if (oauth.enabled && id) {
+        if (oauth.enabled && !isNewUser) {
             fields_.push({
                 name: "oauth_subject",
                 label: oauth.name,
@@ -68,17 +68,15 @@ export function UserWidget({ id }) {
                     widget: PrincipalMemberSelect,
                     choices: group || [],
                     value:
-                        group && id === undefined
-                            ? group
-                                .filter((g) => g.register)
-                                .map((g) => g.id)
+                        group && isNewUser
+                            ? group.filter((g) => g.register).map((g) => g.id)
                             : [],
                 },
                 {
                     name: "language",
                     label: i18n.gettext("Language"),
                     widget: LanguageSelect,
-                    value: 'default',
+                    value: "default",
                 },
                 {
                     name: "description",
@@ -89,12 +87,12 @@ export function UserWidget({ id }) {
         );
 
         return fields_;
-    }, [group]);
+    }, [group, isNewUser]);
 
-    const p = { fields, model: "auth.user", id, messages: getMessages() };
+    const props = { fields, model: "auth.user", id, messages: getMessages() };
 
     // prettier-ignore
-    const infoNGID = useMemo(() => oauth.isNGID && !id && <Alert
+    const infoNGID = useMemo(() => oauth.isNGID && isNewUser && <Alert
         type="info" style={{marginBottom: "1ex"}}
         message={i18n.gettext("Consider adding {name} user to your team instead of creating a new user with a password.").replace("{name}", oauth.name)}
         action={makeTeamManageButton()}
@@ -108,7 +106,7 @@ export function UserWidget({ id }) {
         <div className="ngw-auth-user-widget">
             {infoNGID}
             <ContentBox>
-                <ModelForm {...p}></ModelForm>
+                <ModelForm {...props}></ModelForm>
             </ContentBox>
         </div>
     );