+++ nextgisweb/gui/nodepkg/fields-form/field/Password.js
@@ -1,9 +1,11 @@
 import { Form, Input } from "@nextgisweb/gui/antd";
 
-export function Password({ form, ...props }) {
+export function Password({ form, autoComplete, placeholder, ...props }) {
+    const inputProps = { autoComplete, placeholder };
+
     return (
         <Form.Item {...props}>
-            <Input.Password />
+            <Input.Password {...inputProps} />
         </Form.Item>
     );
 }